src/requirements.txt
beautifulsoup4>=4.9.3
clang>=11.0
fuzzywuzzy>=0.18.0
GitPython>=3.1.13
graphviz>=0.17
imbalanced-learn>=0.8.0
imblearn>=0.0
lxml>=4.6.2
matplotlib-venn>=0.11.6
mypy>=0.800
mysql-connector-python>=8.0.22
numpy>=1.19.4
openpyxl>=3.0.5
pandas>=1.1.5
pyclustering>=0.10.1.2
PyGithub>=1.54.1
pypyodbc>=1.3.4
python-Levenshtein>=0.12.2
requests>=2.25.1
seaborn>=0.11.2
statsmodels>=0.12.2
tensorflow>=2.6.0
types-requests>=2.25.6
xgboost>=1.4.2
xlrd==1.2.0